Added some fucking dumb shit.

+++ commands/end.js
@@ -4,13 +4,18 @@ module.exports = {
     name: 'end',
     description: 'Ends a ticket with a polite conclusion. [STAFF ONLY]',
     execute(message, args) {
-        const embed = new Discord.MessageEmbed()
-            .setAuthor(message.author.tag, message.author.displayAvatarURL())
-            .setDescription(`Thanks for choosing Seaside Creations™ Please leave ${message.author} a review in <#751651872724484137>`)
-        ;
 
-        message.delete();
+        if (message.member.roles.cache.has()) {
+            const embed = new Discord.MessageEmbed()
+                .setAuthor(message.author.tag, message.author.displayAvatarURL())
+                .setDescription(`Thanks for choosing Seaside Creations™ Please leave ${message.author} a review in <#751651872724484137>`)
+            ;
 
-        message.channel.send(embed);
+            message.delete();
+
+            message.channel.send(embed);
+        } else {
+            message.channel.send('No.')
+        }
     }
 }
